requirements.txt
A3MIO==0.0.1
abstract-rendering==0.5.1
airspeed==0.5.5.dev20160812
alabaster==0.7.9
argcomplete==1.8.2
astroid==1.4.9
astropy==1.3
atari-py==0.1.1
attrs==17.2.0
audioread==2.1.4
Automat==0.6.0
Babel==2.3.4
backports-abc==0.5
backports.functools-lru-cache==1.3
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
beautifulsoup4==4.4.1
biopython==1.70
bitarray==0.8.1
blaze==0.9.0
bleach==1.5.0
bokeh==0.12.4
boto==2.45.0
Bottleneck==1.2.0
Box2D-kengz==2.3.3
bunch==1.0.1
cachetools==0.8.0
certifi==2018.4.16
cffi==1.11.5
chardet==2.3.0
chest==0.2.3
click==6.7
cloudpickle==0.2.2
clyent==1.2.1
cmake==0.7.1
colorama==0.3.7
configobj==5.0.6
configparser==3.5.0
constantly==15.1.0
contextlib2==0.5.4
cryptography==1.7.1
cycler==0.10.0
Cython==0.25.2
cytoolz==0.8.2
dask==0.13.0
decorator==4.0.11
dill==0.2.6
Django==1.11
django-jsonify==0.3.0
docutils==0.12
ecdsa==0.13
entrypoints==0.2.2
enum34==1.1.6
et-xmlfile==1.0.1
fastcache==1.0.2
feedparser==5.2.1
filelock==2.0.6
Flask==0.11.1
Flask-Cors==3.0.2
fuel==0.2.0
funcsigs==1.0.2
functools32==3.2.3.post2
futures==3.0.5
gevent==1.1.0
gevent-websocket==0.9.5
greenlet==0.4.12
grin==1.2.1
gym==0.9.4
heapdict==1.0.0
html5lib==0.999
hyperlink==17.2.1
idna==2.1
imageio==2.2.0
imagesize==0.7.1
incremental==17.5.0
ipaddress==1.0.16
ipdb==0.10.3
ipykernel==4.5.2
ipyparallel==6.0.0
ipython==5.2.2
ipython-genutils==0.1.0
ipywidgets==5.2.2
isort==4.2.5
itsdangerous==0.24
jdcal==1.3
jedi==0.10.0
Jinja2==2.8
joblib==0.11
JPype1==0.6.1
jsonschema==2.5.1
jupyter-client==4.4.0
jupyter-console==5.1.0
jupyter-core==4.2.1
Keras==1.2.0
lazy-object-proxy==1.2.2
librosa==0.5.1
line-profiler==2.1.2
llvmlite==0.16.0
locket==0.2.0
lxml==3.7.2
Mako==1.0.6
MarkupSafe==0.23
matplotlib==2.0.0
mccabe==0.6.0
mistune==0.7.3
mock==2.0.0
mpmath==0.19
mujoco-py==0.5.7
multipledispatch==0.4.8
nbconvert==5.1.1
nbformat==4.2.0
networkx==1.11
nltk==3.2.1
nose==1.3.7
notebook==4.4.1
numexpr==2.6.2
numpy==1.13.3
oauthlib==2.0.4
odo==0.5.0
olefile==0.44
openpyxl==2.4.1
pachi-py==0.0.21
pandas==0.19.2
pandocfilters==1.4.1
paramiko==1.18.1
partd==0.3.7
path.py==8.2.1
pathlib2==2.2.1
patsy==0.4.1
pbr==1.10.0
pdbx==0.3.0
Pebble==3.1.15
pep8==1.7.0
PeptideBuilder==1.0.3
pexpect==4.2.1
picklable-itertools==0.1.1
pickleshare==0.7.3
Pillow==4.3.0
pkginfo==1.2.1
ply==3.9
ProDy==1.7
progressbar2==3.10.1
prompt-toolkit==1.0.13
protobuf==3.2.0
psutil==5.1.3
ptyprocess==0.5.1
py==1.4.31
pyasn1==0.1.9
pybullet==1.6.3
pycosat==0.6.3
pycparser==2.17
pycrypto==2.6.1
pyflakes==1.5.0
pyglet==1.2.4
Pygments==2.2.0
PyJWT==1.5.3
pylint==1.6.5
PyOpenGL==3.1.0
PyOpenGL-accelerate==3.1.0
pyOpenSSL==16.2.0
pyparsing==2.1.10
pysvg==0.2.2
PySynth==2.3
pytest==3.0.6
python-dateutil==2.6.0
python-docx==0.8.6
python-openid==2.2.5
python-social-auth==0.3.6
python-utils==2.0.0
pytz==2016.10
PyYAML==3.12
pyzmq==16.0.2
QtAwesome==0.4.4
qtconsole==4.2.1
QtPy==1.2.1
redis==2.10.5
regex==2017.6.23
requests==2.9.1
requests-oauthlib==0.8.0
resampy==0.1.5
rope==0.9.4
ruamel-yaml==0.11.14
scandir==1.3
scikit-image==0.12.3
scikit-learn==0.19.1
scipy==1.1.0
scp==0.10.2
seaborn==0.7.1
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
snowballstemmer==1.2.1
social-auth-app-django==1.2.0
social-auth-core==1.4.0
sockjs-tornado==1.0.3
Sphinx==1.5.2
sphinx-rtd-theme==0.1.9
spyder==2.3.8
SQLAlchemy==1.1.5
statsmodels==0.8.0
subprocess32==3.2.7
sympy==1.0
tensorboard-logger==0.0.3
tensorflow==1.0.0
terminado==0.6
testpath==0.3
Theano==0.9.0
toolz==0.8.2
torch==0.3.1
torchvision==0.2.0
tornado==4.4.2
tqdm==4.9.0
traitlets==4.3.1
Twisted==17.5.0
typing==3.6.4
ujson==1.35
unicodecsv==0.14.1
wcwidth==0.1.7
Werkzeug==0.11.10
widgetsnbextension==1.2.6
wikipedia==1.4.0
wokkel==0.7.1
wrapt==1.10.8
xlrd==1.0.0
XlsxWriter==0.9.6
xlwt==1.2.0
zope.interface==4.4.2
